chore(spec): drop conversions of values already uint64
The YAML structs hold the gas limits as uint64, so converting the field is
a no-op that unconvert flags.

+++ spec/gloas/executionpayloadbid_yaml.go
@@ -51,7 +51,7 @@ func (e *ExecutionPayloadBid) MarshalYAML() ([]byte, error) {
 		BlockHash:             fmt.Sprintf("%#x", e.BlockHash),
 		PrevRandao:            fmt.Sprintf("%#x", e.PrevRandao),
 		FeeRecipient:          fmt.Sprintf("%#x", e.FeeRecipient),
-		GasLimit:              uint64(e.GasLimit),
+		GasLimit:              e.GasLimit,
 		BuilderIndex:          uint64(e.BuilderIndex),
 		Slot:                  uint64(e.Slot),
 		Value:                 uint64(e.Value),
